refactor: move kite actions hook to specifc file
Move kite actions hook to specifc file following the pattern in other
components.
